requirements_new.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
anyio==3.6.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
biopython==1.79
bleach==5.0.1
brendapy==0.5.0
cached-property==1.5.2
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
ConfigArgParse==1.5.3
connection-pool==0.0.3
coverage==6.4.4
cycler==0.11.0
datrie==0.8.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
depinfo==2.1.1
distlib==0.3.5
docutils==0.17.1
dpath==2.0.6
entrypoints==0.4
ete3==3.1.2
executing==1.0.0
fastjsonschema==2.15.3
fastobo==0.12.2
filelock==3.7.1
fonttools==4.37.1
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==3.2.0
jupyter-book==0.13.1
jupyter-cache==0.4.3
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.18.1
jupyter-server-mathjax==0.2.6
jupyter-sphinx==0.3.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
latexcodec==2.0.1
linkify-it-py==1.0.3
lxml==4.9.2
markdown-it-py==1.1.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.2.8
mistune==0.8.4
multidict==6.0.2
myst-nb==0.13.2
myst-parser==0.15.2
nbclient==0.5.13
nbconvert==6.5.2
nbdime==3.1.1
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.12
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plac==1.3.5
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
pronto==2.5.0
psutil==5.9.1
ptyprocess==0.7.0
PuLP==2.6.0
pure-eval==0.2.2
py==1.11.0
py3Dmol==1.8.1
pybtex==0.24.0
pybtex-docutils==1.0.2
pycparser==2.21
pydantic==1.9.1
pydata-sphinx-theme==0.8.1
Pygments==2.12.0
pymetadata==0.3.3
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
pytest==7.1.3
pytest-cov==4.0.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
ratelimiter==1.2.0.post0
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
retry==0.9.2
rich==12.5.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
smart-open==6.0.0
smmap==5.0.0
snakemake==7.8.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==4.5.0
sphinx-book-theme==0.3.3
sphinx-comments==0.0.3
sphinx-copybutton==0.5.0
sphinx-external-toc==0.2.4
sphinx-jupyterbook-latex==0.4.6
sphinx-multitoc-numbering==0.1.3
sphinx-thebe==0.1.2
sphinx-togglebutton==0.3.2
sphinx_design==0.1.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.40
stack-data==0.5.0
stopit==1.1.2
tabulate==0.8.9
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toposort==1.7
tornado==6.2
tox==3.25.1
traitlets==5.3.0
typing_extensions==4.3.0
uc-micro-py==1.0.1
ujson==5.4.0
urllib3==1.26.9
uuid==1.30
virtualenv==20.15.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
widgetsnbextension==3.6.1
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.7.2
yte==1.4.0
zeep==4.1.0
zipp==3.8.1
